requirements.txt
Flask==2.3.2
requests==2.32.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytz==2024.1
Werkzeug==3.1.2
cloudinary==1.33.0
